refactor: move the epochs related constants
